# [GraphBolt][CUDA] Optimize UVA `index_select_csc`.