## Description
Inside the UVA read kernel, we avoid expensive binary search operations by performing it earlier and passing the result as a parameter. This ensures that the only expensive operation inside the actual UVA read kernel is the UVA reads, it should help with resource usage. (help make the 6144 default enough for this kernel as well.)

<details><summary>Before the PR</summary>
<p>

```
* params:  n_rows=128000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 22153, Throughput in MiB/s: 2066


* params:  n_rows=128000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 1942, Throughput in MiB/s: 2356


* params:  n_rows=128000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 164, Throughput in MiB/s: 278


* params:  n_rows=128000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 21677, Throughput in MiB/s: 2111


* params:  n_rows=128000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 1975, Throughput in MiB/s: 2317


* params:  n_rows=128000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 227, Throughput in MiB/s: 201


* params:  n_rows=16000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 17665, Throughput in MiB/s: 14683


* params:  n_rows=16000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 1841, Throughput in MiB/s: 14087


* params:  n_rows=16000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 155, Throughput in MiB/s: 1664


* params:  n_rows=16000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 15793, Throughput in MiB/s: 16424


* params:  n_rows=16000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 1681, Throughput in MiB/s: 15422


* params:  n_rows=16000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 213, Throughput in MiB/s: 1214


* params:  n_rows=16000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 10004, Throughput in MiB/s: 4574


* params:  n_rows=16000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 1107, Throughput in MiB/s: 4130


* params:  n_rows=16000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 163, Throughput in MiB/s: 279


* params:  n_rows=16000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 9508, Throughput in MiB/s: 4814


* params:  n_rows=16000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 1053, Throughput in MiB/s: 4346


* params:  n_rows=16000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 217, Throughput in MiB/s: 210


* params:  n_rows=16000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 10803, Throughput in MiB/s: 7061


* params:  n_rows=16000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 1200, Throughput in MiB/s: 6355


* params:  n_rows=16000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 169, Throughput in MiB/s: 451


* params:  n_rows=16000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 9460, Throughput in MiB/s: 8064


* params:  n_rows=16000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 1096, Throughput in MiB/s: 6957


* params:  n_rows=16000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 170, Throughput in MiB/s: 447


* params:  n_rows=2000000, avg_degree=256, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 47301, Throughput in MiB/s: 20969


* params:  n_rows=2000000, avg_degree=256, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 4982, Throughput in MiB/s: 19902


* params:  n_rows=2000000, avg_degree=256, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 189, Throughput in MiB/s: 5212


* params:  n_rows=2000000, avg_degree=256, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 45596, Throughput in MiB/s: 21752


* params:  n_rows=2000000, avg_degree=256, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 4755, Throughput in MiB/s: 20857


* params:  n_rows=2000000, avg_degree=256, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 282, Throughput in MiB/s: 3505


* params:  n_rows=2000000, avg_degree=256, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 89906, Throughput in MiB/s: 21890


* params:  n_rows=2000000, avg_degree=256, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 8998, Throughput in MiB/s: 21878


* params:  n_rows=2000000, avg_degree=256, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 194, Throughput in MiB/s: 10124


* params:  n_rows=2000000, avg_degree=256, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 87095, Throughput in MiB/s: 22600


* params:  n_rows=2000000, avg_degree=256, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 8801, Throughput in MiB/s: 22364


* params:  n_rows=2000000, avg_degree=256, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 196, Throughput in MiB/s: 10033


* params:  n_rows=2000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 16274, Throughput in MiB/s: 15944


* params:  n_rows=2000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 1876, Throughput in MiB/s: 13826


* params:  n_rows=2000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 232, Throughput in MiB/s: 1116


* params:  n_rows=2000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 14777, Throughput in MiB/s: 17553


* params:  n_rows=2000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 1640, Throughput in MiB/s: 15809


* params:  n_rows=2000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 157, Throughput in MiB/s: 1644


* params:  n_rows=2000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 25380, Throughput in MiB/s: 19836


* params:  n_rows=2000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 2843, Throughput in MiB/s: 17700


* params:  n_rows=2000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 156, Throughput in MiB/s: 3225


* params:  n_rows=2000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 24303, Throughput in MiB/s: 20718


* params:  n_rows=2000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 2683, Throughput in MiB/s: 18760


* params:  n_rows=2000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 159, Throughput in MiB/s: 3153


* params:  n_rows=2000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 8734, Throughput in MiB/s: 5240


* params:  n_rows=2000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 1065, Throughput in MiB/s: 4297


* params:  n_rows=2000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 201, Throughput in MiB/s: 227


* params:  n_rows=2000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 8524, Throughput in MiB/s: 5369


* params:  n_rows=2000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 1045, Throughput in MiB/s: 4377


* params:  n_rows=2000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 379, Throughput in MiB/s: 120


* params:  n_rows=2000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 9255, Throughput in MiB/s: 8243


* params:  n_rows=2000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 1133, Throughput in MiB/s: 6730


* params:  n_rows=2000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 160, Throughput in MiB/s: 476


* params:  n_rows=2000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 8439, Throughput in MiB/s: 9039


* params:  n_rows=2000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 1037, Throughput in MiB/s: 7350


* params:  n_rows=2000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int64, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 222, Throughput in MiB/s: 342


Total runtimes in us:  {<Device.Pinned: 1>: 559405}
```

</p>
</details> 

<details><summary>After PR</summary>
<p>

```
* params:  n_rows=128000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 22641, Throughput in MiB/s: 2021


* params:  n_rows=128000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 1965, Throughput in MiB/s: 2329


* params:  n_rows=128000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 233, Throughput in MiB/s: 195


* params:  n_rows=128000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 22641, Throughput in MiB/s: 2021


* params:  n_rows=128000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 2052, Throughput in MiB/s: 2230


* params:  n_rows=128000000, avg_degree=8, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.REGULAR

Runtime in us: 213, Throughput in MiB/s: 214


* params:  n_rows=16000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 19365, Throughput in MiB/s: 13398


* params:  n_rows=16000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=100000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 2163, Throughput in MiB/s: 11988


* params:  n_rows=16000000, avg_degree=64, indptr_dtype=torch.int64, tensor_dtype=torch.int32, num_indices=1000, indptr_device=Device.Pinned, tensor_device=Device.Pinned, indices_device=Device.GPU, regular=Regularity.IRREGULAR

Runtime in us: 215, Throughput in MiB/s: 1201

…