@mfbalin 2024-01-23
> Can you quantify the optimization?

@frozenbugs I updated the description. Ran the benchmark test with and without limiting the number of cuda threads used for the UVA copy. It looks like there is some regression when we don't limit the number of cuda threads used, 559ms -> 600ms but when we limit the CUDA threads to 6144, 1028ms -> 702ms as expected. So my optimization is doing what it promises, I will see if I can reduce the amount of regression for the unlimited CUDA thread case.