@mfbalin 2024-01-23
The general case is bound to regress a little bit because instead of having these kernels fused as before and having the UVA copy start immediately, now we launch one more kernel before performing the actual copy operation, so the PCI-e copies start later, which means they can utilize the PCI-e slightly less with unlimited number of threads.